refactor: add healthcheck domain
