+++ app/src/main/java/com/codesoom/assignment/healthcheck/controller/HelloController.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.controllers;
+package com.codesoom.assignment.healthcheck.controller;
 
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
